requirements.txt
APScheduler==3.3.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pbr==3.1.1
pytz==2017.2
requests==2.18.4
six==1.11.0
slackclient==1.0.9
SQLAlchemy==1.1.14
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
tzlocal==1.4
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
WTForms==2.1
gunicorn==19.1.1
psycopg2==2.5.4

